+++ src/views/Header/HeaderTitle.js
@@ -15,7 +15,7 @@ const HeaderTitle = ({ style, ...rest }) => (
 const styles = StyleSheet.create({
   title: {
     fontSize: Platform.OS === 'ios' ? 17 : 20,
-    fontWeight: Platform.OS === 'ios' ? '700' : '500',
+    fontWeight: Platform.OS === 'ios' ? '600' : '500',
     color: 'rgba(0, 0, 0, .9)',
     marginHorizontal: 16,
   },